Extracted InMemoryLedgerStateAccess.
